+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
   - hardis:work:save : call forceSourcePull method to propose to update .forceignore if errors are found
   - hardis:project:configure:auth: call mdapi:deploy with RunLocalTests to manage production environments
   - authentication: auth only to devHub if --devhub sent
+  - Disable spinner for restDeploy check
 
 ## [2.8.5] 2021-04-06